requirements.txt
fastapi
uvicorn[standard]
pydantic
chromadb
openai
streamlit
python-dotenv
requests
pytest
pytest-asyncio
pytest-mock